Implemented a default tab setting.
When handling tab decrease, we shouldn't assume the user wants to remove all indentation at the start of the line. With a new 'default tab' setting, we can define the tab setting (number of spaces) for the editor. Now, when a user wants to decrease their indentation, it will do it one tab at a time.